[Add-on, Fix] Architecture - see desc
- Add a closure around workerScript functions in order to avoid
overwriting from user functions/globals.

- Added FakeWorker constructor as a compatibility fallback for browsers
not supporting workers.

- Removed Thread constructor (since it is unnecessary now due to the
new FakeWorker) and moved its functionality to ThreadInterface.

- Created a new helpers object containing helper functions and moved
all stringify/parse methods to it.

- Improve worker support test by attempting to create a sample worker
thread and immediately terminating it inside a try/catch block which
will resolve to false if it throws. This correctly marks browsers that
support workers but not workers from blob URIs (modernIE).

- Included new helpers object as a global dependency for thread workers
as a replacement to manual declaration of stringify/parse functions.